requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0
tensorflow-lite>=2.12.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# DCT and Transforms
jpeg2dct>=0.1.0
pytorch-dct>=0.0.1

# Data Processing
pandas>=2.0.0
h5py>=3.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Metrics and Evaluation
scikit-learn>=1.3.0
lpips>=0.1.4  # Perceptual similarity
pytorch-fid>=0.3.0  # FID score

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0

# Utilities
tqdm>=4.65.0
rich>=13.0.0  # Beautiful terminal output
wandb>=0.15.0  # Optional: experiment tracking

# Model Optimization
onnx>=1.14.0
onnxruntime>=1.15.0
torch-pruning>=1.2.0

# Mobile Deployment
coremltools>=7.0  # iOS deployment
tf2onnx>=1.14.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Dataset Management
kaggle>=1.5.16
gdown>=4.7.1  # Google Drive downloads

# Security and Cryptography (for secret message handling)
cryptography>=41.0.0
pycryptodome>=3.18.0

# Performance Profiling
py-spy>=0.3.14
memory-profiler>=0.61.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Jupyter Support (optional)
jupyter>=1.0.0
ipywidgets>=8.1.0
